Split Contracts page into components and styles
Refactor admin Contracts UI by extracting inline CSS and large page logic into reusable components and a stylesheet. Adds components: ContractCreateModal, ContractEditModal, ContractStatusBadge, ContractStatusModal, ContractViewModal, ContractsStats, ContractsTable, ContractsToolbar and helpers (contractsHelpers.js), plus AdminContracts.css. Updates src/pages/admin/Contracts.jsx to import the new components and stylesheet and remove the big inline styles/duplicated helpers, simplifying the page and improving modularity and reuse.
